Fix BearSSLClient profile configuration
